more number conversion changes (#1096)
* more number conversion changes

* try to fix tests

* Update poi-integration-exceptions.csv

+++ poi/src/main/java/org/apache/poi/ss/format/SimpleFraction.java
@@ -133,9 +133,9 @@ public class SimpleFraction {
         }
 
         if (q2 < maxDenominator) {
-            return new SimpleFraction((int) p2, (int)q2);
+            return new SimpleFraction(Math.toIntExact(p2), Math.toIntExact(q2));
         } else {
-            return new SimpleFraction((int)p1, (int)q1);
+            return new SimpleFraction(Math.toIntExact(p1), Math.toIntExact(q1));
         }
 
     }
